Add Playwright to run on preview environments
this adds playwright to run on the preview environments

it creates a reusable step ofr running the e2e.
we removed the defualt playwright github action, and
integrated that into areuasble step.

There is now a test that runs a simple test on the
preview environment, that the homepage loads.

The result is uplaoded as an artifact as a zip file.
ideally that woudl be uplaoded to s3 so you can just lcikc on it